refactor(users): adicionar tratamento de erro para usuário não encontrado

+++ backend/src/users/users.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersRepository } from './repositories/users.repository';
+import { NotFoundError } from 'src/common/errors/types/NotFoundError';
 
 @Injectable()
 export class UsersService {
@@ -12,10 +13,22 @@ export class UsersService {
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const user = await this.repository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundError('User not found');
+    }
+
     return this.repository.update(id, updateUserDto);
   }
 
   async remove(id: number) {
+    const user = await this.repository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundError('User not exist');
+    }
+
     return this.repository.remove(id);
   }
 
@@ -24,6 +37,12 @@ export class UsersService {
   }
 
   async findOne(id: number) {
-    return this.repository.findOne(id);
+    const user = await this.repository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundError('User not found');
+    }
+
+    return user;
   }
 }
